开发代码 0.根据新加入的服务，重新修补了“删除”类接口的业务逻辑
by organwalk 2023-10-22

+++ training-department-service/src/main/java/com/training/department/service/Impl/DepartmentServiceImpl.java
@@ -5,6 +5,7 @@ import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.training.common.entity.*;
 import com.training.common.entity.req.UserInfoListReq;
+import com.training.department.client.ResourceClient;
 import com.training.department.client.UserClient;
 import com.training.department.entity.request.DeptReq;
 import com.training.department.entity.request.MembersReq;
@@ -30,6 +31,7 @@ import java.util.Objects;
 public class DepartmentServiceImpl implements DepartmentService {
 
     private final UserClient userClient;
+    private final ResourceClient resourceClient;
     private final DeptMapper deptMapper;
     private final DeptWorkerMapper deptWorkerMapper;
 
@@ -115,6 +117,11 @@ public class DepartmentServiceImpl implements DepartmentService {
         Integer deptMark = deptMapper.selectDeptExist(deptId);
         if (Objects.isNull(deptMark)){
             return MsgRespond.fail("该部门不存在，无法删除");
+        }
+        // 检查此部门是否设置了部门标签
+        Integer codeMark = resourceClient.getTagListByDeptId(deptId).getInteger("code");
+        if (Objects.equals(codeMark, 2002)){
+            return MsgRespond.fail("该部门已定义资源分类标签，无法删除。请在删除资源分类标签后重试");
         }
         // 先删除指定部门下的员工
         deptWorkerMapper.deleteAllWorkerByDeptId(deptId);
